services/search/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
weaviate-client==4.6.3
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1